Update to headers from JB 4.1.2
This was mostly done since this was tested on a device with JB 4.1.2,
it's unknown if the ABI of these APIs are stable across 4.0-4.3
like the normal IOMX code. If this actually is the case, one would
have to test how many of the API versions in 4.0-4.3 actually
require different headers.